+++ packages/grafana-llm-app/pkg/plugin/azure_provider.go
@@ -1,27 +1,41 @@
 package plugin
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/url"
 	"time"
+
+	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
+	"github.com/sashabaranov/go-openai"
 )
 
 type azure struct {
-	settings OpenAISettings
-	c        *http.Client
+	settings     OpenAISettings
+	defaultModel Model
+	oc           *openai.Client
 }
 
-func NewAzureProvider(settings OpenAISettings) LLMProvider {
-	return &azure{
-		settings: settings,
-		c: &http.Client{
-			Timeout: 2 * time.Minute,
-		},
+func NewAzureProvider(settings OpenAISettings, defaultModel Model) (LLMProvider, error) {
+	client := &http.Client{
+		Timeout: 2 * time.Minute,
 	}
+	p := &azure{
+		settings:     settings,
+		defaultModel: defaultModel,
+	}
+
+	// go-openai expects the URL without the '/openai' suffix, which is
+	// the same as us.
+	cfg := openai.DefaultAzureConfig(settings.apiKey, settings.URL)
+	cfg.HTTPClient = client
+	// We pass the deployment as the name of the model, so just return the untransformed string.
+	cfg.AzureModelMapperFunc = func(model string) string {
+		return model
+	}
+
+	p.oc = openai.NewClientWithConfig(cfg)
+	return p, nil
 }
 
 func (p *azure) Models(ctx context.Context) (ModelResponse, error) {
@@ -36,41 +50,47 @@ func (p *azure) Models(ctx context.Context) (ModelResponse, error) {
 	return ModelResponse{Data: models}, nil
 }
 
-type azureChatCompletionRequest struct {
-	ChatCompletionRequest
-	// Azure does not use the model field.
-	Model string `json:"-"`
-}
-
-func (p *azure) ChatCompletions(ctx context.Context, req ChatCompletionRequest) (ChatCompletionsResponse, error) {
+func (p *azure) getDeployment(model Model) (string, error) {
 	mapping, err := p.getAzureMapping()
 	if err != nil {
-		return ChatCompletionsResponse{}, err
+		return "", err
 	}
-	deployment := mapping[req.Model]
+	if model == "" {
+		model = p.defaultModel
+	}
+	deployment := mapping[model]
 	if deployment == "" {
-		return ChatCompletionsResponse{}, fmt.Errorf("%w: no deployment found for model: %s", errBadRequest, req.Model)
+		return "", fmt.Errorf("%w: no deployment found for model: %s", errBadRequest, model)
+	}
+	return deployment, nil
+}
+
+func (p *azure) ChatCompletion(ctx context.Context, req ChatCompletionRequest) (openai.ChatCompletionResponse, error) {
+	deployment, err := p.getDeployment(req.Model)
+	if err != nil {
+		return openai.ChatCompletionResponse{}, err
 	}
 
-	u, err := url.Parse(p.settings.URL)
+	r := req.ChatCompletionRequest
+	r.Model = deployment
+	resp, err := p.oc.CreateChatCompletion(ctx, r)
 	if err != nil {
-		return ChatCompletionsResponse{}, err
+		log.DefaultLogger.Error("error creating azure chat completion", "err", err)
+		return openai.ChatCompletionResponse{}, err
 	}
-	u.Path = fmt.Sprintf("/openai/deployments/%s/chat/completions", deployment)
-	u.RawQuery = "api-version=2023-03-15-preview"
-	reqBody, err := json.Marshal(azureChatCompletionRequest{
-		ChatCompletionRequest: req,
-		Model:                 req.Model.toOpenAI(),
-	})
+	return resp, nil
+}
+
+func (p *azure) ChatCompletionStream(ctx context.Context, req ChatCompletionRequest) (<-chan ChatCompletionStreamResponse, error) {
+	deployment, err := p.getDeployment(req.Model)
 	if err != nil {
-		return ChatCompletionsResponse{}, err
+		return nil, err
 	}
-	httpReq, err := http.NewRequestWithContext(ctx, http.MethodPost, u.String(), bytes.NewReader(reqBody))
-	if err != nil {
-		return ChatCompletionsResponse{}, err
-	}
-	httpReq.Header.Set("api-key", p.settings.apiKey)
-	return doOpenAIRequest(p.c, httpReq)
+
+	r := req.ChatCompletionRequest
+	// For the Azure mapping we want to use the name of the mapped deployment as the model.
+	r.Model = deployment
+	return streamOpenAIRequest(ctx, r, p.oc)
 }
 
 func (p *azure) getAzureMapping() (map[Model]string, error) {