+++ packages/grafana-llm-app/pkg/plugin/resources_test.go
@@ -170,8 +170,18 @@ func newMockOpenAIServer(t *testing.T) *mockServer {
 	server := &mockServer{}
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		server.request = r
+		streaming := r.Header.Get("Accept") == "text/event-stream"
+		if streaming {
+			w.Header().Set("Content-Type", "text/event-stream")
+			w.WriteHeader(http.StatusOK)
+			w.Write([]byte("data: {}\n\n"))
+			w.Write([]byte("data: [DONE]\n\n"))
+			return
+		}
 		w.WriteHeader(http.StatusOK)
+		w.Header().Set("Content-Type", "application/json")
 		w.Write([]byte("{}"))
+
 	})
 	server.server = httptest.NewServer(handler)
 	return server
@@ -224,6 +234,86 @@ func TestCallOpenAIProxy(t *testing.T) {
 			expStatus: http.StatusOK,
 		},
 		{
+			name: "openai - empty model",
+
+			settings: Settings{
+				OpenAI: OpenAISettings{
+					OrganizationID: "myOrg",
+					Provider:       openAIProviderOpenAI,
+				},
+			},
+			apiKey: "abcd1234",
+
+			method: http.MethodPost,
+			path:   "/openai/v1/chat/completions",
+			body:   []byte(`{"messages": [{"content":"some stuff"}]}`),
+
+			expReqHeaders: http.Header{
+				"Authorization":       {"Bearer abcd1234"},
+				"OpenAI-Organization": {"myOrg"},
+			},
+			expReqPath: "/v1/chat/completions",
+			expReqBody: []byte(`{"model": "gpt-3.5-turbo", "messages": [{"content":"some stuff"}]}`),
+
+			expStatus: http.StatusOK,
+		},
+		{
+			name: "openai - streaming",
+
+			settings: Settings{
+				OpenAI: OpenAISettings{
+					OrganizationID: "myOrg",
+					Provider:       openAIProviderOpenAI,
+				},
+			},
+			apiKey: "abcd1234",
+
+			method: http.MethodPost,
+			path:   "/openai/v1/chat/completions",
+			body:   []byte(`{"model": "base", "stream": true, "messages": [{"content":"some stuff"}]}`),
+
+			expReqHeaders: http.Header{
+				"Authorization":       {"Bearer abcd1234"},
+				"OpenAI-Organization": {"myOrg"},
+			},
+			expReqPath: "/v1/chat/completions",
+			expReqBody: []byte(`{"model": "gpt-3.5-turbo", "stream": true, "messages": [{"content":"some stuff"}]}`),
+
+			expStatus: http.StatusOK,
+
+			// We need to use regular strings rather than raw strings here otherwise the double
+			// newlines (required by the SSE spec) are escaped.
+			expBody: []byte("data: {\"id\":\"\",\"object\":\"\",\"created\":0,\"model\":\"\",\"choices\":null,\"system_fingerprint\":\"\"}\n\ndata: [DONE]\n\n"),
+		},
+		{
+			name: "openai - streaming - empty model",
+
+			settings: Settings{
+				OpenAI: OpenAISettings{
+					OrganizationID: "myOrg",
+					Provider:       openAIProviderOpenAI,
+				},
+			},
+			apiKey: "abcd1234",
+
+			method: http.MethodPost,
+			path:   "/openai/v1/chat/completions",
+			body:   []byte(`{"stream": true, "messages": [{"content":"some stuff"}]}`),
+
+			expReqHeaders: http.Header{
+				"Authorization":       {"Bearer abcd1234"},
+				"OpenAI-Organization": {"myOrg"},
+			},
+			expReqPath: "/v1/chat/completions",
+			expReqBody: []byte(`{"model": "gpt-3.5-turbo", "stream": true, "messages": [{"content":"some stuff"}]}`),
+
+			expStatus: http.StatusOK,
+
+			// We need to use regular strings rather than raw strings here otherwise the double
+			// newlines (required by the SSE spec) are escaped.
+			expBody: []byte("data: {\"id\":\"\",\"object\":\"\",\"created\":0,\"model\":\"\",\"choices\":null,\"system_fingerprint\":\"\"}\n\ndata: [DONE]\n\n"),
+		},
+		{
 			name: "azure",
 
 			settings: Settings{
@@ -252,6 +342,62 @@ func TestCallOpenAIProxy(t *testing.T) {
 			expStatus: http.StatusOK,
 		},
 		{
+			name: "azure - abstract model",
+
+			settings: Settings{
+				OpenAI: OpenAISettings{
+					OrganizationID: "myOrg",
+					Provider:       openAIProviderAzure,
+					AzureMapping: [][]string{
+						{"gpt-3.5-turbo", "gpt-35-turbo"},
+					},
+				},
+			},
+
+			apiKey: "abcd1234",
+
+			method: http.MethodPost,
+			path:   "/openai/v1/chat/completions",
+			body:   []byte(`{"model": "base", "messages": [{"content":"some stuff"}]}`),
+
+			expReqHeaders: http.Header{
+				"api-key": {"abcd1234"},
+			},
+			expReqPath: "/openai/deployments/gpt-35-turbo/chat/completions",
+			// the 'model' field should have been removed.
+			expReqBody: []byte(`{"messages":[{"content":"some stuff"}]}`),
+
+			expStatus: http.StatusOK,
+		},
+		{
+			name: "azure - empty model",
+
+			settings: Settings{
+				OpenAI: OpenAISettings{
+					OrganizationID: "myOrg",
+					Provider:       openAIProviderAzure,
+					AzureMapping: [][]string{
+						{"gpt-3.5-turbo", "gpt-35-turbo"},
+					},
+				},
+			},
+
+			apiKey: "abcd1234",
+
+			method: http.MethodPost,
+			path:   "/openai/v1/chat/completions",
+			body:   []byte(`{"messages": [{"content":"some stuff"}]}`),
+
+			expReqHeaders: http.Header{
+				"api-key": {"abcd1234"},
+			},
+			expReqPath: "/openai/deployments/gpt-35-turbo/chat/completions",
+			// the 'model' field should have been removed.
+			expReqBody: []byte(`{"messages":[{"content":"some stuff"}]}`),
+
+			expStatus: http.StatusOK,
+		},
+		{
 			name: "azure invalid deployment",
 
 			settings: Settings{
@@ -268,7 +414,7 @@ func TestCallOpenAIProxy(t *testing.T) {
 			method: http.MethodPost,
 			path:   "/openai/v1/chat/completions",
 			// note no gpt-4 in AzureMapping.
-			body: []byte(`{"model": "gpt-4", "messages": [{"content":"some stuff"}]}`),
+			body: []byte(`{"model": "gpt-4-turbo", "messages": [{"content":"some stuff"}]}`),
 
 			expNilRequest: true,
 
@@ -291,7 +437,7 @@ func TestCallOpenAIProxy(t *testing.T) {
 			body:   []byte(`{"model": "gpt-3.5-turbo", "messages": [{"content":"some stuff"}]}`),
 
 			expReqHeaders: http.Header{
-				"Authorization": {"Basic MTIzOmFiY2QxMjM0"},
+				"Authorization": {"Bearer 123:abcd1234"},
 				"X-Scope-OrgID": {"123"},
 			},
 			expReqPath: "/llm/openai/v1/chat/completions",
@@ -316,7 +462,32 @@ func TestCallOpenAIProxy(t *testing.T) {
 			body:   []byte(`{"model": "gpt-3.5-turbo", "messages": [{"content":"some stuff"}]}`),
 
 			expReqHeaders: http.Header{
-				"Authorization": {"Basic MTIzOmFiY2QxMjM0"},
+				"Authorization": {"Bearer 123:abcd1234"},
+				"X-Scope-OrgID": {"123"},
+			},
+			expReqPath: "/llm/openai/v1/chat/completions",
+			expReqBody: []byte(`{"model": "gpt-3.5-turbo", "messages": [{"content":"some stuff"]}}`),
+
+			expStatus: http.StatusOK,
+		},
+		{
+			name: "grafana-managed llm gateway - empty model",
+
+			settings: Settings{
+				Tenant:           "123",
+				GrafanaComAPIKey: "abcd1234",
+				OpenAI: OpenAISettings{
+					Provider: openAIProviderGrafana,
+				},
+			},
+			apiKey: "abcd1234",
+
+			method: http.MethodPost,
+			path:   "/openai/v1/chat/completions",
+			body:   []byte(`{"messages": [{"content":"some stuff"}]}`),
+
+			expReqHeaders: http.Header{
+				"Authorization": {"Bearer 123:abcd1234"},
 				"X-Scope-OrgID": {"123"},
 			},
 			expReqPath: "/llm/openai/v1/chat/completions",
@@ -362,6 +533,8 @@ func TestCallOpenAIProxy(t *testing.T) {
 				t.Fatal("inst must be of type *App")
 			}
 
+			app.ignoreResponsePadding = true
+
 			var r mockCallResourceResponseSender
 			err = app.CallResource(ctx, &backend.CallResourceRequest{
 				PluginContext: backend.PluginContext{
@@ -400,8 +573,8 @@ func TestCallOpenAIProxy(t *testing.T) {
 				t.Errorf("response status should be %d, got %d", tc.expStatus, r.response.Status)
 			}
 			if len(tc.expBody) > 0 {
-				if tb := bytes.TrimSpace(r.response.Body); !bytes.Equal(tb, tc.expBody) {
-					t.Errorf("response body should be %s, got %s", tc.expBody, tb)
+				if !bytes.Equal(r.response.Body, tc.expBody) {
+					t.Errorf("response body should be %s, got %s", tc.expBody, r.response.Body)
 				}
 			}
 		})