+++ packages/grafana-llm-frontend/src/openai.ts
@@ -72,27 +72,29 @@ export interface Function {
   parameters: Object;
 }
 
+/**
+ * Enum representing abstracted models used by the backend app.
+ * @enum {string}
+ */
+export enum Model {
+  BASE = 'base',
+  LARGE = 'large',
+}
+
+/**
+ * @deprecated Use {@link Model} instead.
+ */
+type DeprecatedString = string;
+
 export interface ChatCompletionsRequest {
   /**
-   * ID of the model to use.
+   * Model abstraction to use. These abstractions are then translated back into specific models based on the users settings.
    *
-   * See the model endpoint compatibility table for details on which models work with the Chat Completions API.
+   * If not specified, defaults to `Model.BASE`.
    */
-  model: string;
+  model?: Model | DeprecatedString;
   /** A list of messages comprising the conversation so far. */
   messages: Message[];
-  /** A list of functions the model may generate JSON inputs for. */
-  functions?: Function[];
-  /**
-   * Controls how the model responds to function calls.
-   *
-   * "none" means the model does not call a function, and responds to the end-user.
-   * "auto" means the model can pick between an end-user or calling a function.
-   * Specifying a particular function via {"name": "my_function"} forces the model to call that function.
-   *
-   * "none" is the default when no functions are present. "auto" is the default if functions are present.
-   */
-  function_call?: 'none' | 'auto' | { name: string };
   /**
    * What sampling temperature to use, between 0 and 2.
    * Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic.
@@ -255,7 +257,7 @@ export function isErrorResponse<T>(
  * @returns An observable that emits the content messages. Each emission will be a string containing the
  *         token emitted by the model.
  * @example <caption>Example of reading all tokens in a stream.</caption>
- * const stream = streamChatCompletions({ model: 'gpt-3.5-turbo', messages: [
+ * const stream = streamChatCompletions({ model: Model.BASE, messages: [
  *   { role: 'system', content: 'You are a great bot.' },
  *   { role: 'user', content: 'Hello, bot.' },
  * ]}).pipe(extractContent());
@@ -282,7 +284,7 @@ export function extractContent(): UnaryFunction<
  * @returns An observable that emits the accumulated content messages. Each emission will be a string containing the
  *         content of all messages received so far.
  * @example
- * const stream = streamChatCompletions({ model: 'gpt-3.5-turbo', messages: [
+ * const stream = streamChatCompletions({ model: Model.BASE, messages: [
  *   { role: 'system', content: 'You are a great bot.' },
  *   { role: 'user', content: 'Hello, bot.' },
  * ]}).pipe(accumulateContent());
@@ -324,7 +326,7 @@ export async function chatCompletions(request: ChatCompletionsRequest): Promise<
  * The 'done' message will not be emitted; the stream will simply end when this message is encountered.
  *
  * @example <caption>Example of reading all tokens in a stream.</caption>
- * const stream = streamChatCompletions({ model: 'gpt-3.5-turbo', messages: [
+ * const stream = streamChatCompletions({ model: Model.BASE, messages: [
  *   { role: 'system', content: 'You are a great bot.' },
  *   { role: 'user', content: 'Hello, bot.' },
  * ]}).pipe(extractContent());
@@ -333,7 +335,7 @@ export async function chatCompletions(request: ChatCompletionsRequest): Promise<
  * // ['Hello', '? ', 'How ', 'are ', 'you', '?']
  *
  * @example <caption>Example of accumulating tokens in a stream.</caption>
- * const stream = streamChatCompletions({ model: 'gpt-3.5-turbo', messages: [
+ * const stream = streamChatCompletions({ model: Model.BASE, messages: [
  *   { role: 'system', content: 'You are a great bot.' },
  *   { role: 'user', content: 'Hello, bot.' },
  * ]}).pipe(accumulateContent());
@@ -487,7 +489,7 @@ export type OpenAIStreamState = {
  * @property {Subscription|undefined} value.stream - The stream subscription object if the stream is active, or undefined if not.
  */
 export function useOpenAIStream(
-  model = 'gpt-4',
+  model = Model.LARGE,
   temperature = 1,
   notifyError: (title: string, text?: string, traceId?: string) => void = () => {}
 ): OpenAIStreamState {